Implement description editing, for now
